% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_meta.R
\name{concordantGtypesMeta}
\alias{concordantGtypesMeta}
\title{See if a gtypes and a meta object have concordant information}
\usage{
concordantGtypesMeta(meta, gin)
}
\arguments{
\item{meta}{meta data frame}

\item{gin}{gtypes object (must have populations as strata).}
}
\value{
boolean T if all required cols are present in meta
}
\description{
Tests to see if the populations in meta match up to the strata in gin
}
