% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introModels.R
\name{equalSourceTbl}
\alias{equalSourceTbl}
\title{shipping no longer than 30 days from sources}
\usage{
equalSourceTbl(intros = NULL, sources = NULL)
}
\arguments{
\item{intros}{vector of names of intro regions}

\item{sources}{vector of names of source regions}
}
\description{
shipping no longer than 30 days from sources
}
