% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcRF_model.R
\name{errRFmodel}
\alias{errRFmodel}
\title{Calculate Error for ABC Random Forest Model}
\usage{
errRFmodel(rfFit, refobj, modelColumn = "introModel", cores = 1)
}
\arguments{
\item{rfFit}{A fitted ABC Random Forest model.}

\item{refobj}{A reference object.}

\item{modelColumn}{A string specifying the model column. Default is "introModel".}

\item{cores}{An integer specifying the number of cores to use. Default is 1.}
}
\value{
Error metrics for the ABC Random Forest model.
}
\description{
This function calculates the error of an ABC Random Forest model.
}
