% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_priors.R
\name{getPriors}
\alias{getPriors}
\title{defines prior distributions and generates a pull for each call}
\usage{
getPriors()
}
\value{
a list of prior values
}
\description{
The parameters that are estimated are each chosen from prior distributions in this function.
}
