% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introModels.R
\name{gigasTblAdmix}
\alias{gigasTblAdmix}
\title{gigas admixture-based introduction model}
\usage{
gigasTblAdmix(intros = NULL, sources = NULL)
}
\arguments{
\item{intros}{vector of names of intro regions}

\item{sources}{vector of names of source regions}
}
\description{
gigas admixture-based introduction model
}
