% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introModels.R
\name{gigasTblML}
\alias{gigasTblML}
\title{gigas machine learning model based introduction table}
\usage{
gigasTblML(intros = NULL, sources = NULL)
}
\arguments{
\item{intros}{vector of names of intro regions}

\item{sources}{vector of names of source regions}
}
\description{
gigas machine learning model based introduction table
}
