% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainparam.R
\name{mainparams}
\alias{mainparams}
\title{creates the species and problem-level parameters}
\usage{
mainparams(
  cores = 1,
  nreps = 1,
  species = "",
  demeN = 0,
  demeI = 0,
  sources = c("kag", "hok", "tok", "hon", "sea"),
  intros = c("PNW", "Cali", "NZ", "Arg", "EU"),
  dataType = "",
  fsc_exec = "",
  exclude = "nonSource"
)
}
\arguments{
\item{cores}{number of cores (usually set to 1)}

\item{nreps}{number of reps to run per computation}

\item{species}{name of species under consideration}

\item{demeN}{controls ghost pops in native regions}

\item{demeI}{controls ghost pops in introduced regions}

\item{sources}{the possible source regions for this species (must have a sample in each)}

\item{intros}{the introduced regions (must have a sample in each of these regions)}

\item{dataType}{the type of genetic data (found in datafile.R)}

\item{fsc_exec}{name of the fastsimcoal executable}

\item{exclude}{source populations to exclude from analysis (ususally nonsource)}
}
\value{
list containing parameters
}
\description{
The sources and intros are calculated when you run the
    'species_setup()' function and when the simulation hierarchy is
    created, these are stored in a file called datafiles.R in the
    'src'subdirectory.  This function is really only used in the call to setupReps()
}
