% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{make_pca_ref_table}
\alias{make_pca_ref_table}
\title{make_pca_ref_table}
\usage{
make_pca_ref_table(refObj, prop.variation = 0.99, scale.var = T)
}
\arguments{
\item{refObj}{An object containing reference data. TODO: Add more detailed description.}

\item{prop.variation}{Proportion of variation to be captured. Default is 0.99.}

\item{scale.var}{Logical. Whether to scale the variables. Default is TRUE.}
}
\value{
A table resulting from the PCA. TODO: Add more detailed description.
}
\description{
Perform PCA on reference data and return a table.
}
\examples{
# TODO: Add example usage for make_pca_ref_table
}
