% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildref.R
\name{make_ref_table}
\alias{make_ref_table}
\title{make_ref_table}
\usage{
make_ref_table(
  genofile,
  mname,
  intros,
  sources,
  csvpath = ".",
  dataType = "microsatellite",
  na.prop = 0.01,
  scale.var = T,
  popPairwise = F,
  verb = F
)
}
\arguments{
\item{genofile}{Path to the genetic data file. TODO: Add more detailed description.}

\item{mname}{TODO: Describe this parameter.}

\item{intros}{TODO: Describe this parameter.}

\item{sources}{TODO: Describe this parameter.}

\item{csvpath}{Path to save the CSV output. Default is '.'.}

\item{dataType}{Type of genetic data. Default is 'microsatellite'.}

\item{na.prop}{Proportion of NAs. Default is 0.01.}

\item{scale.var}{Logical. Whether to scale the variables. Default is TRUE.}

\item{verb}{Logical. Whether to print details (default FALSE)}
}
\value{
A reference table based on genetic data. TODO: Add more detailed description.
}
\description{
Create a reference table based on genetic data. TODO: Add more detailed description.
}
\examples{
# TODO: Add example usage for make_ref_table
}
