% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcRF_model.R
\name{modelRF}
\alias{modelRF}
\title{Build an ABC Random Forest Model}
\usage{
modelRF(refobj, modelColumn = "introModel", cores = 1, ...)
}
\arguments{
\item{refobj}{A reference object.}

\item{modelColumn}{A string specifying the model column. Default is "introModel".}

\item{cores}{An integer specifying the number of cores to use. Default is 1.}

\item{...}{Additional arguments.}
}
\value{
An ABC Random Forest model.
}
\description{
This function builds an Approximate Bayesian Computation (ABC) Random Forest model.
}
