% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeModels.R
\name{nativeHistory}
\alias{nativeHistory}
\title{set up the history for the native range}
\usage{
nativeHistory(
  sources = c("kag", "hon", "sea", "tok", "hok"),
  topology = matrix(c(0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 3, 2, 0, 0, 0, 0,
    0, 0, 0), byrow = T, nrow = 5, dimnames = list(c("kag", "hon", "sea", "tok", "hok"),
    c("kag", "hon", "sea", "tok", "hok")))
)
}
\arguments{
\item{sources}{vector of source regions (should be same as rownames for intro tables)

the native range topology maps the times t1-t3 to the 4 regions in japan
will use this in the generate coalescence model functions to determine within
native range coalescence.  This works by going down columns of the matrix.  Hokkaido
coalesces into honshu at time 1 and honshu into tokyo at time 3 (at least for the
original gvermSNP file.  if this is downstream, regions likely differ}
}
\description{
set up the history for the native range
}
