% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoseSumStatsModel.R
\name{obsModelDist}
\alias{obsModelDist}
\title{obsModelDist}
\usage{
obsModelDist(refObj, plot = FALSE)
}
\arguments{
\item{refObj}{An object containing reference data. TODO: Add more detailed description.}

\item{plot}{Logical. If TRUE, plots the differences. Default is FALSE.}
}
\value{
A vector of distances. TODO: Add more detailed description.
}
\description{
Calculate the distance between observed and model distributions.
}
