% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramEstimate.R
\name{paramPosteriors}
\alias{paramPosteriors}
\title{paramPosteriors function}
\usage{
paramPosteriors(
  SimObj,
  cores = 1,
  method = "neuralnet",
  tol = c(0.1, 0.05, 0.01, 0.005, 0.001, 0.005)
)
}
\arguments{
\item{SimObj}{A simulation object containing reference data, parameters, and observed data.}

\item{cores}{Number of cores to use for parallel computation. Default is 1.}

\item{method}{Method to use for ABC. Default is "neuralnet".}

\item{tol}{Tolerance levels for ABC. Default is c(0.1,0.05,0.01,0.005,0.001,0.005).}
}
\value{
A list of ABC results for each combination of method and tolerance level.
Each element of the list is named by concatenating the method and tolerance level with an underscore.
}
\description{
This function computes the posterior distribution of parameters using Approximate Bayesian Computation (ABC).
}
\examples{
# Assuming appropriate data and functions are available:
result <- paramPosteriors(SimObj)

}
