% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPca.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(
  pca.ref,
  prop_points = 0.2,
  mod_color = c(2:5),
  obs_color = 6,
  models = c("Ship30", "Ship60", "gigasAdmix", "gigasML")
)
}
\arguments{
\item{pca.ref}{A reference for the PCA results. TODO: Add more detailed description.}

\item{prop_points}{Proportion of points to be displayed. Default is 0.2.}

\item{mod_color}{Color for the model. Default is c(2:5).}
}
\value{
A plot visualizing the PCA results. TODO: Add more detailed description.
}
\description{
Plot the results of a PCA.
}
\examples{
# TODO: Add example usage for plotPCA
}
