% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcRF_model.R
\name{plotRFmodel}
\alias{plotRFmodel}
\title{Plot ABC Random Forest Model}
\usage{
plotRFmodel(rfFit, refobj, modelColumn = "introModel")
}
\arguments{
\item{rfFit}{A fitted ABC Random Forest model.}

\item{refobj}{A reference object.}

\item{modelColumn}{A string specifying the model column. Default is "introModel".}
}
\value{
Plots related to the ABC Random Forest model.
}
\description{
This function plots an ABC Random Forest model.
}
