% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcRF_model.R
\name{ref2RF}
\alias{ref2RF}
\title{ref2RF Utility Function}
\usage{
ref2RF(refobj, modelColumn)
}
\arguments{
\item{refobj}{A reference object.}

\item{modelColumn}{A string specifying the model column.}
}
\value{
A data frame prepared for Random Forest.
}
\description{
This internal function converts reference objects for Random Forest.
}
\keyword{internal}
