% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqgen2gtypes.R
\name{seqgen2gtype}
\alias{seqgen2gtype}
\title{Convert take a tree create sequences using phyclust::seqgen and convert to gtypes}
\usage{
seqgen2gtype(tree, params, temp.file, demes, mu = 1)
}
\arguments{
\item{tree}{A phylogenetic tree to base sequence generation on.}

\item{params}{A list containing parameters for sequence generation.}

\item{temp.file}{A temporary file location to write sequences to.}

\item{demes}{A list or data frame containing information about demes or populations.}

\item{mu}{A numeric parameter for mutation rate. Default is 1.0.}
}
\value{
A gtypes object.
}
\description{
This function generates sequences based on a given phylogenetic tree and parameters 
using the seqgen function from the phyclust package. It then converts these sequences 
into gtype format using the strataG::sequence2gtypes function.
}
