% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupReps.R
\name{setupReps}
\alias{setupReps}
\title{setup a single simulation rep for abc}
\usage{
setupReps(mname, genofile, params = mainparams())
}
\arguments{
\item{mname}{name of metadata file for pops}

\item{genofile}{name of gtypes file}

\item{params}{a list of parameters that come from mainparams()}
}
\value{
a list of parameters needed for coalescent simulation creation
}
\description{
This function takes the meta data and the geentic data that are placed
in the 'data' subfolder to create a set of parameters that are common to all simulation
replicates.  This function only need to be called once when a set of simulations starts.
}
