% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_setup.R
\name{species_setup}
\alias{species_setup}
\title{Set up the file hierarchy and data files}
\usage{
species_setup(
  root = "test",
  fas = NULL,
  indmeta = NULL,
  genofile = NULL,
  mname = NULL,
  newdir = NULL,
  dataType = "sequence",
  species = "test",
  nativeTopology = NULL,
  fsc_exec = "fsc27",
  popPairwise = FALSE
)
}
\arguments{
\item{root}{name of the root directory in the hierarchy to get information}

\item{fas}{name of a fasta file containing haplotypes. will be converted to a gtypes}

\item{genofile}{filename of a gtypes object in .RDS format (must have populations as strata). Either fas or genofilehave to be specified}

\item{mname}{name of file containing metadata that links pops to strata}

\item{newdir}{path to the new directory to set things up for a species}

\item{dataType}{type of data ("sequence", "snp", "microsatellite")}

\item{species}{the name of the species under consideration}

\item{nativeTopology}{a named matrix describing the topology and relative times for the history of a species in the native range.  If null defaults to output of function nativeHistory()}

\item{fsc_exec}{executable for fastsimcoal}
}
\description{
Set up the file hierarchy and data files
}
