% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats_seq}
\alias{summary_stats_seq}
\title{calculate summary stats for sequence data (like mtDNA or cpDNA)}
\usage{
summary_stats_seq(gin, meta)
}
\arguments{
\item{gin}{is a gtypes object with sequence data}

\item{meta}{is a dataframe with metadata}
}
\description{
calculates a lot of population genetic summary statistics
}
