\name{testInvPath-package}
\alias{testInvPath-package}
\alias{testInvPath}
\docType{package}
\title{
\packageTitle{testInvPath}
}
\description{
\packageDescription{testInvPath}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testInvPath}
\packageIndices{testInvPath}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{testInvPath}

Maintainer: \packageMaintainer{testInvPath}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
